sample-gpt-api-call/pyproject.toml
[project]
name = "sample-gpt-api-call"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.3.0",
    "python-dotenv>=1.1.1",
]
